requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
crcmod==1.7
idna==3.4
Nuitka==1.3.5
ordered-set==4.1.0
requests==2.28.1
urllib3==1.26.13
zstandard==0.19.0
google-re2==1.0
